chore: settle ground to move sprites to ryot_sprites
