+++ apps/ryot_compass/src/lib.rs
@@ -58,6 +58,7 @@ impl Plugin for AppPlugin {
                 DefaultPlugins
                     .set(entitled_window("Compass".to_string()))
                     .set(ImagePlugin::default_nearest()),
+                RyotLegacySpritePlugin,
                 VisualContentPlugin::<CompassContentAssets>::default(),
                 WorldInspectorPlugin::default()
                     .run_if(action_toggle_active(false, ToggleFeatures::Inspector)),
