+++ crates/ryot/src/bevy_ryot/mod.rs
@@ -3,20 +3,12 @@
 //! This module is intended to be used as a library dependency for RyOT games.
 //! It provides common ways of dealing with OT content, such as loading sprites,
 //! configuring the game, and handling asynchronous events.
-use self::sprites::SpriteMaterial;
 #[cfg(feature = "debug")]
 use crate::position::debug_sprite_position;
 use crate::prelude::*;
 use bevy::app::{App, Plugin, Update};
-use bevy::asset::embedded_asset;
 use bevy::prelude::*;
-use bevy::sprite::Material2dPlugin;
-use bevy_asset_loader::loading_state::LoadingStateAppExt;
-use bevy_asset_loader::prelude::*;
-use bevy_asset_loader::standard_dynamic_asset::StandardDynamicAssetArrayCollection;
-use bevy_common_assets::json::JsonAssetPlugin;
 use bevy_stroked_text::StrokedTextPlugin;
-use std::marker::PhantomData;
 
 mod game;
 pub use game::*;
@@ -31,38 +23,13 @@ pub mod sprites;
 pub(crate) mod sprite_animations;
 pub use sprite_animations::{toggle_sprite_animation, AnimationDuration};
 
-content_plugin!(VisualContentPlugin, VisualElementsAsset);
+pub struct RyotLegacySpritePlugin;
 
-impl<C> Plugin for VisualContentPlugin<C>
-where
-    C: VisualElementsAsset + CatalogAsset + AtlasLayoutsAsset + Default,
-{
+impl Plugin for RyotLegacySpritePlugin {
     fn build(&self, app: &mut App) {
-        app.add_plugins(BaseContentPlugin::<C>::default())
-            .configure_loading_state(
-                LoadingStateConfig::new(RyotContentState::LoadingContent)
-                    .with_dynamic_assets_file::<StandardDynamicAssetArrayCollection>(
-                    "dynamic.atlases.ron",
-                ),
-            )
-            .init_resource::<RectMeshes>()
-            .init_resource::<SpriteMeshes>()
-            .init_resource::<SpriteSheetDataSet>()
-            .init_resource::<TextureAtlasLayouts>()
-            .add_plugins(JsonAssetPlugin::<Catalog>::new(&["json"]))
-            .add_optional_plugin(StrokedTextPlugin)
-            .add_plugins(Material2dPlugin::<SpriteMaterial>::default())
-            .add_systems(
-                OnEnter(RyotContentState::PreparingContent),
-                (
-                    prepare_sprite_layouts::<C>,
-                    prepare_sprite_sheets::<C>,
-                    prepare_sprite_meshes,
-                    transition_to_ready,
-                )
-                    .chain()
-                    .after(prepare_visual_elements::<C>),
-            )
+        app.add_plugins(ryot_sprites::prelude::RyotSpritePlugin);
+
+        app.add_optional_plugin(StrokedTextPlugin)
             .init_resource::<sprite_animations::SpriteAnimationEnabled>()
             .init_resource::<sprite_animations::SynchronizedAnimationTimers>()
             .init_resource::<sprites::LoadedAppearances>()
@@ -98,20 +65,5 @@ where
                     (move_sprites_with_animation, finish_position_animation).chain(),
                 ),
             );
-
-        embedded_asset!(app, "shaders/sprite.wgsl");
     }
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Clone, SystemSet)]
-pub enum SpriteSystems {
-    Load,
-    Initialize,
-    Update,
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Clone, SystemSet)]
-pub enum AnimationSystems {
-    Initialize,
-    Update,
 }
