+++ crates/ryot_content/src/game/mod.rs
@@ -1,0 +1,1 @@
+pub mod game_object;
