+++ crates/ryot_content/src/sprite/layout.rs
