+++ crates/ryot_sprites/src/material/shaders/sprite.wgsl
