--- crates/ryot_sprites/src/sheet/mod.rs
@@ -1,1 +1,0 @@
-pub mod definitions;
